user input for bin width in preprocess command
