+++ wrappers/preprocess_wrap.py
@@ -3,7 +3,7 @@ import logging
 logger = logging.getLogger(__name__)
 
 
-def preprocess_wrap(fasta_dir, output_dir, output_base_qual, chunk_size, threads):
+def preprocess_wrap(fasta_dir, output_dir, output_base_qual, chunk_size, bin_size, threads):
     import os
     import time
     import resource
@@ -22,7 +22,7 @@ def preprocess_wrap(fasta_dir, output_dir, output_base_qual, chunk_size, threads
     if len(files_to_process) == 1:
         # If there is only one file, process it in a single thread
         logger.info("Only one file to process. Sorting reads into bins without parallelization.")
-        extract_and_bin_reads(files_to_process[0], chunk_size, output_dir + "/full_length_pp_fa", output_base_qual)
+        extract_and_bin_reads(files_to_process[0], chunk_size, output_dir + "/full_length_pp_fa", output_base_qual, bin_size)
         os.system(f"rm {output_dir}/full_length_pp_fa/*.lock")
 
         logger.info("Converting tsv files into parquet")
@@ -32,7 +32,7 @@ def preprocess_wrap(fasta_dir, output_dir, output_base_qual, chunk_size, threads
         # Process files in parallel
         logger.info("Multiple raw files found. Sorting reads into bins with parallelization.")
         parallel_preprocess_data(
-            files_to_process, output_dir + "/full_length_pp_fa", chunk_size, output_base_qual, num_workers=threads
+            files_to_process, output_dir + "/full_length_pp_fa", chunk_size, output_base_qual, bin_size, num_workers=threads
         )
     usage = resource.getrusage(resource.RUSAGE_CHILDREN)
     max_rss_mb = usage.ru_maxrss / 1024 if os.uname().sysname == "Linux" else usage.ru_maxrss
